# Optimize stripcslashes