@fabpot 2024-08-12
> @fabpot Could you sync this back to v4 so that I can continue?

I've just merged 3.x into 4.x.